fix missile not moving on create when no ship
If the Saucer and missile are created when the ship happens to be
destroyed the missile will be drawn but will not move.  This is because
the base.Move() method is called with a 0 X and Y velocity set.  This
changes it to call the Missile.Move overload and provides a randomized
point target so it does not simply move in a straight line.